pyproject.toml
[tool.poetry]
name = "promptgit"
version = "0.1.0"
description = "Managing prompts for LLM in Git repository"
authors = ["Boris Tvaroska <boris@tvaroska.sk>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pydantic = ">=2.5.2,<3.0"
GitPython = ">=3.0.0,<4.0.0"
urllib3 = ">=2.2.2,<3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^24.3.0"
isort = "^5.13.2"
tox = "^4.11.4"
pylint = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80